Move WickedRenkoConsolidator to RenkoConsolidator as Default (#5931)
* Rename Renko to ClassicRenko

* Rename WickedRenko to Renko

* Adjust tests and usages

* Backwards support WickedRenkoConsolidator

* Renames for consistency
